requirements.txt
asgiref==3.8.1
certifi==2025.4.26
Django==5.2.1
django-filter==25.1
django-jazzmin==3.0.1
django-jet==1.0.8
djangorestframework==3.16.0
jplephem==2.22
numpy==2.2.5
pillow==11.2.1
psycopg2-binary==2.9.10
pytz==2023.3
sgp4==2.24
skyfield==1.53
sqlparse==0.5.3
